refactor: 将 modelPlatforms 从 config/ 迁移至 utils/ 以统一工具函数目录
